Renames and moves assembler classes
Bean assemblers moved into the o.e.s.e.queues.spooler.beanassemblers
package. Model assemblers moved into the
o.e.s.queues.spooler.modelassmblers package. Also renames
IScanPathModelAssembler to IScanObjectAssembler.
