+++ org.eclipse.scanning.event.queues/src/org/eclipse/scanning/event/queues/spooler/beanassemblers/MonitorAtomAssembler.java
@@ -1,4 +1,4 @@
-package org.eclipse.scanning.event.queues.spooler;
+package org.eclipse.scanning.event.queues.spooler.beanassemblers;
 
 import org.eclipse.scanning.api.event.queues.IQueueBeanFactory;
 import org.eclipse.scanning.api.event.queues.beans.MonitorAtom;
