+++ org.eclipse.scanning.event.queues/src/org/eclipse/scanning/event/queues/spooler/beanassemblers/PositionerAtomAssembler.java
@@ -1,4 +1,4 @@
-package org.eclipse.scanning.event.queues.spooler;
+package org.eclipse.scanning.event.queues.spooler.beanassemblers;
 
 import java.util.stream.Collectors;
