+++ org.eclipse.scanning.event.queues/src/org/eclipse/scanning/event/queues/spooler/modelassemblers/AbstractPathModelAssembler.java
@@ -1,11 +1,11 @@
-package org.eclipse.scanning.event.queues.spooler.pathassemblers;
+package org.eclipse.scanning.event.queues.spooler.modelassemblers;
 
 import java.util.Arrays;
 import java.util.List;
 
 import org.eclipse.scanning.api.points.models.IScanPathModel;
 
-public abstract class AbstractPathModelAssembler<P extends IScanPathModel> implements IScanPathModelAssembler<P> {
+public abstract class AbstractPathModelAssembler<P extends IScanPathModel> implements IScanObjectModelAssembler<P> {
 
 	protected final String[] required;
