+++ org.eclipse.scanning.event.queues/src/org/eclipse/scanning/event/queues/spooler/modelassemblers/IScanObjectModelAssembler.java
@@ -1,11 +1,11 @@
-package org.eclipse.scanning.event.queues.spooler.pathassemblers;
+package org.eclipse.scanning.event.queues.spooler.modelassemblers;
 
 import java.util.List;
 
 import org.eclipse.scanning.api.event.queues.models.DeviceModel;
 import org.eclipse.scanning.api.event.queues.models.QueueModelException;
 
-public interface IScanPathModelAssembler<P> {
+public interface IScanObjectModelAssembler<P> {
 	
 	P assemble(String name, DeviceModel model) throws QueueModelException;
